Restructure some stuff, add babel stuff to get env to build
