+++ package.json
@@ -25,14 +25,17 @@
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject",
     "storybook": "start-storybook -p 9009 -s public",
-    "build-storybook": "build-storybook -s public"
+    "build-storybook": "build-storybook -s public",
+    "build-cli": "babel src/logic --out-dir build/logic",
+    "montage": "node build/logic/cli"
   },
   "devDependencies": {
     "@storybook/addon-actions": "^3.3.13",
     "@storybook/addon-links": "^3.3.13",
     "@storybook/addons": "^3.3.13",
     "@storybook/react": "^3.3.13",
-    "babel-core": "^6.26.0"
+    "babel-core": "^6.26.0",
+    "babel-preset-env": "^1.6.1"
   },
   "quokka": {
     "env": {
